# modules/bike/handlers.go
package bike

import (
	"net/http"

	"github.com/MauroRaya/bike-rental-api/httputil"
	repo "github.com/MauroRaya/bike-rental-api/sqlc/autogenerated"
)

type handler struct {
	service Service
}

type Handler interface {
	ListBikes(w http.ResponseWriter, r *http.Request)
	FindBikeByID(w http.ResponseWriter, r *http.Request)
	CreateBike(w http.ResponseWriter, r *http.Request)
	UpdateBike(w http.ResponseWriter, r *http.Request)
	DeleteBike(w http.ResponseWriter, r *http.Request)
}

func NewHandler(service Service) Handler {
	return &handler{service}
}

// ListBikes godoc
// @Summary List bikes
// @Description Get all bikes
// @Tags bike
// @Produce json
// @Success 200 {array} repo.Bike
// @Failure 500 {string} string
// @Router /bike [get]
func (h *handler) ListBikes(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	bikes, err := h.service.ListBikes(ctx)
	if err != nil {
		http.Error(w, "unexpected error", http.StatusInternalServerError)
		return
	}

	httputil.EncodeJSON(w, http.StatusOK, bikes)
}

// FindBikeByID godoc
// @Summary Find bike by ID
// @Description Get a bike by ID
// @Tags bike
// @Produce json
// @Param id path int true "Bike ID"
// @Success 200 {object} repo.Bike
// @Failure 422 {string} string
// @Failure 500 {string} string
// @Router /bike/{id} [get]
func (h *handler) FindBikeByID(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	id, err := httputil.ParsePathID(r)
	if err != nil {
		http.Error(w, "invalid id", http.StatusUnprocessableEntity)
		return
	}

	bike, err := h.service.FindBikeByID(ctx, id)
	if err != nil {
		http.Error(w, "unexpected error", http.StatusInternalServerError)
		return
	}

	httputil.EncodeJSON(w, http.StatusOK, bike)
}

// CreateBike godoc
// @Summary Create bike
// @Description Create a new bike
// @Tags bike
// @Accept json
// @Produce json
// @Param payload body repo.CreateBikeParams true "Bike payload"
// @Success 201 {object} repo.Bike
// @Failure 422 {string} string
// @Failure 500 {string} string
// @Router /bike [post]
func (h *handler) CreateBike(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var payload repo.CreateBikeParams

	if err := httputil.DecodeJSON(r, &payload); err != nil {
		http.Error(w, "invalid request", http.StatusUnprocessableEntity)
		return
	}

	bike, err := h.service.CreateBike(ctx, payload)
	if err != nil {
		http.Error(w, "unexpected error", http.StatusInternalServerError)
		return
	}

	httputil.EncodeJSON(w, http.StatusCreated, bike)
}

// UpdateBike godoc
// @Summary Update bike
// @Description Update a bike
// @Tags bike
// @Accept json
// @Produce json
// @Param id path int true "Bike ID"
// @Param payload body repo.UpdateBikeParams true "Bike payload"
// @Success 200 {object} repo.Bike
// @Failure 422 {string} string
// @Failure 500 {string} string
// @Router /bike/{id} [put]
func (h *handler) UpdateBike(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var payload repo.UpdateBikeParams

	if err := httputil.DecodeJSON(r, &payload); err != nil {
		http.Error(w, "invalid request", http.StatusUnprocessableEntity)
		return
	}

	bike, err := h.service.UpdateBike(ctx, payload)
	if err != nil {
		http.Error(w, "unexpected error", http.StatusInternalServerError)
		return
	}

	httputil.EncodeJSON(w, http.StatusOK, bike)
}

// UpdateBike godoc
// @Summary Update bike
// @Description Update a bike
// @Tags bike
// @Accept json
// @Produce json
// @Param id path int true "Bike ID"
// @Param payload body repo.UpdateBikeParams true "Bike payload"
// @Success 200 {object} repo.Bike
// @Failure 422 {string} string
// @Failure 500 {string} string
// @Router /bike/{id} [put]
func (h *handler) DeleteBike(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	id, err := httputil.ParsePathID(r)
	if err != nil {
		http.Error(w, "invalid id", http.StatusUnprocessableEntity)
		return
	}

	bike, err := h.service.DeleteBike(ctx, id)
	if err != nil {
		http.Error(w, "unexpected error", http.StatusInternalServerError)
		return
	}

	httputil.EncodeJSON(w, http.StatusOK, bike)
}